Remove not needed checks on collections
There is no need to check if an element is not yet in a set
before adding it. Set collection protects us against duplicates so there
is no need to explicitely check for it.

Similarily there is no need to check if an element is in a list before
trying to remove it. If the element is in the list, checking for its
presense before removing it will only add not needed processing.